aggiornamenti su acquisto dentro un ordine

+++ app/src/main/java/com/fabio/autenticazione/DTO/ProdottoAcquistoDTO.java
@@ -1,5 +1,5 @@
 package com.fabio.autenticazione.DTO;
 
-public record ProdottoAcquistoDTO(String nome, int quantita, int totale,int prezzoUnitario) {
+public record ProdottoAcquistoDTO(String nome, int quantita, Double totale,Double prezzoUnitario) {
     
 }
